Add files via upload
This program calculates the RSI's (Relative Strength Index) of four stocks as a momentum indicator. It then uses this data to train a logistic regression model to predict when to buy or sell and then these signals are graphed for better visualisation. This code also involves a backtesting section that is commented for easier readability.

Error: When calculating the RSI, the values are either too large or too small. This error is a W.I.P.